Speed up image resizer (by huibao@) If the origin returns non html content while flushing early, trigger a redirect to a noscript version of the page with flush early disabled (by poojatandon@)

+++ src/pagespeed/kernel/image/image_resizer.h
@@ -110,7 +110,7 @@ class ScanlineResizer : public ScanlineReaderInterface {
   int row_;
 
   // Buffer for storing the intermediate results.
-  scoped_array<uint8> buffer_;
+  scoped_array<float> buffer_;
   int bytes_per_buffer_row_;
 
   DISALLOW_COPY_AND_ASSIGN(ScanlineResizer);
